Implements an efficient way to provide random numbers for radiation
NOTE: BFB with different number of procs but NBFB restarts.
This commit recodes the way random numbers are sent to radiation codes.
The existing method was very inefficient. To improve it, now each
column is assigned a column id. This information is passed on to each
task. Seeds are generated using column ids. These seeds are then used
to produce random numbers.

[NBFB]