Infrastructure for incorporating dyn_npes_stride into SE dycore
This commit changes the paradigm under which an individual MPI
task identified as a dynamics processor thus allowing for the
incorporation of the dyn_npes_stride namelist variable into
the SE dycore.

If dyn_npes_stride = N then the MPI tasks that will be assigned
as dynamics cores will be every N tasks up to the total number
of dyn_npes tasks.

This distrubution provides speed up in the solution even in the
naive implementation where the physics columns are still distributed
over all of the compute nodes.

[BFB] - Bit-For-Bit