bugfix kludge: allow neg interp time in find_times_to_interpolate
In volc_rad_data, time interpolation to a target time < the 1st source
time would use the time of the LAST source time as the left edge of the
interpolation interval, causing weird extrapolation. This fix sets that
left edge to that last time - 365 days so interp works correctly. This
is NOT a permanent solution because it assumes the time units are days
and that a 365 day calendar is used.

modified:   cam/src/chemistry/utils/volc_rad_data.F90